requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pydantic==2.6.3
textblob==0.17.1
nltk==3.8.1
matplotlib==3.8.3
pandas==2.2.1
python-dotenv==1.0.1
pytest==8.0.0
httpx==0.27.0
